I did a dummy map and removed the object-oriented code.
I did a couple of dummy sprites and filled the entire map area with a dummy sprite. When I tried to run the cartridge, though, the cartridge had trouble with some of the object-oriented code. I commented out the problematic part, but the cartridge also had some errors in the conditionals of the actual code. I fixed those, but then the bject-oriented code caused a runtime error. While thinking about that problem, I realized that this game, a simple maze-exploration game, had no need for object-oriented code. I took out the object-oriented code, but I plan on getting the code to work with the Pico-8 because I need the functionality for my other game: Re-Hoard.